--- Costaline/Costaline/Custom/Slot.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Costaline
-{   
-    public class Slot
-    {
-        public string name;
-        Domain _domain;
-        public string value;
-
-        public void AddDomains(Domain domain)
-        {
-            _domain = domain;       
-        }
-    }
-}
