--- Costaline/Costaline/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="GraphX" version="2.3.6.0" targetFramework="net472" />
-  <package id="QuickGraphPCL" version="3.6.61114.2" targetFramework="net472" />
-</packages>
