+++ Costaline/Custom/Domain.cs
@@ -6,15 +6,15 @@ using System.Threading.Tasks;
 
 namespace Costaline
 {
-    class Domain
+    public class Domain
     {
-        public string Name { get; set; }// ToDo сделать более осмысленое имя. Пока оно определяет для какого слота используется. Напиться.
-        public List<string> Values { get; set; }
+        public string name { get; set; }// ToDo сделать более осмысленое имя. Пока оно определяет для какого слота используется. Напиться.
+        public List<string> values { get; set; }
 
         public Domain()
         {
-            Values = new List<string>();
+            values = new List<string>();
         }
-        // Возможно стоит добавить добавление доменов
+        // не должно быть возможности добавить домен
     }
 }
