+++ Costaline/Custom/MLV.cs
@@ -8,6 +8,17 @@ namespace Costaline
 {
     class MLV
     {
+        List<Frame> _sistemFrames;
+        Finder _finder = new Finder();
 
+        public void LoadSistemFrames(List<Frame> frames)
+        {
+            _sistemFrames = frames;
+        }
+
+        public Slot Getanswer(Frame frame)
+        {
+            return null;
+        }
     }
 }
